+++ tests/test-files/jest/nok.spec.ts
@@ -1,0 +1,16 @@
+describe('getNumberEnumKeys', () => {
+
+    beforeEach(() => {/* A hook */});
+    beforeAll(() => {/* A hook */});
+    afterEach(() => {/* A hook */});
+    afterAll(() => {/* A hook */});
+
+    it('zero-indexed enum', () => {
+        const values = getNumberEnumKeys(MyNumberEnum);
+        expect(values).toStrictEqual([
+            'ValueZero',
+            'ValueOne',
+            'ValueTwo',
+        ]);
+    });
+});