+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[comment]: <> (## [Unreleased])
+## [Unreleased]
+### Added
+- Support for eslint-plugin-sonarjs@^0.6.0
+- Support for eslint-plugin-sonarjs@^0.7.0
+
 ## [3.0.0] - 2021-01-10
 ### BREAKING CHANGES
 - Upgraded to @typescript-eslint/eslint-plugin 4.12.0.