+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to my ESLint config will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+[comment]: <> (## [Unreleased])
+## [4.0.0] - 2021-07-27
 ### BREAKING CHANGES
 - Added eslint-plugin-jest to peerDependencies.
 - Added eslint-plugin-unicorn to peerDependencies.
@@ -64,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial version.
 
-[Unreleased]: https://github.com/Ionaru/eslint-config/compare/3.1.0...HEAD
+[Unreleased]: https://github.com/Ionaru/eslint-config/compare/4.0.0...HEAD
+[4.0.0]: https://github.com/Ionaru/eslint-config/compare/3.1.0...4.0.0
 [3.1.0]: https://github.com/Ionaru/eslint-config/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/Ionaru/eslint-config/compare/2.0.0...3.0.0
 [2.0.0]: https://github.com/Ionaru/eslint-config/compare/1.0.4...2.0.0